Switch from likes to number of times searched
It didn't make sense to add a like to recommendation items since there aren't any additional information about the recommendation items besides the name. If a like button were to be added, a new document would need to be created for each recommended show. Then, similar items would have to be added to each recommended item (probably if the item was searched for the field would update). This might be implemented in the future, but for now likes will be changed to number of times searched.

Note: to keep search count inflation, perhaps do something with unique visitors.